Move stuff that affects the whole screen/viewport out of Drawing.{cpp,h}
