+++ src/openrct2-ui/windows/DebugPaint.cpp
@@ -12,6 +12,7 @@
 #include <openrct2-ui/windows/Windows.h>
 #include <openrct2/Context.h>
 #include <openrct2/core/Guard.hpp>
+#include <openrct2/drawing/Drawing.Screen.h>
 #include <openrct2/drawing/Drawing.String.h>
 #include <openrct2/drawing/Drawing.h>
 #include <openrct2/drawing/Font.h>
@@ -76,37 +77,37 @@ namespace OpenRCT2::Ui::Windows
             {
                 case WIDX_TOGGLE_SHOW_WIDE_PATHS:
                     gPaintWidePathsAsGhost = !gPaintWidePathsAsGhost;
-                    GfxInvalidateScreen();
+                    Drawing::GfxInvalidateScreen();
                     break;
 
                 case WIDX_TOGGLE_SHOW_BLOCKED_TILES:
                     gPaintBlockedTiles = !gPaintBlockedTiles;
-                    GfxInvalidateScreen();
+                    Drawing::GfxInvalidateScreen();
                     break;
 
                 case WIDX_TOGGLE_SHOW_SEGMENT_HEIGHTS:
                     gShowSupportSegmentHeights = !gShowSupportSegmentHeights;
-                    GfxInvalidateScreen();
+                    Drawing::GfxInvalidateScreen();
                     break;
 
                 case WIDX_TOGGLE_SHOW_BOUND_BOXES:
                     gPaintBoundingBoxes = !gPaintBoundingBoxes;
-                    GfxInvalidateScreen();
+                    Drawing::GfxInvalidateScreen();
                     break;
 
                 case WIDX_TOGGLE_SHOW_DIRTY_VISUALS:
                     gShowDirtyVisuals = !gShowDirtyVisuals;
-                    GfxInvalidateScreen();
+                    Drawing::GfxInvalidateScreen();
                     break;
 
                 case WIDX_TOGGLE_STABLE_PAINT_SORT:
                     gPaintStableSort = !gPaintStableSort;
-                    GfxInvalidateScreen();
+                    Drawing::GfxInvalidateScreen();
                     break;
 
                 case WIDX_TOGGLE_FORCE_REDRAW:
                     gPaintForceRedraw = !gPaintForceRedraw;
-                    GfxInvalidateScreen();
+                    Drawing::GfxInvalidateScreen();
                     break;
             }
         }
