+++ src/openrct2/actions/peep/StaffSetNameAction.cpp
@@ -12,7 +12,7 @@
 #include "../../Context.h"
 #include "../../Diagnostic.h"
 #include "../../GameState.h"
-#include "../../drawing/Drawing.h"
+#include "../../drawing/Drawing.Screen.h"
 #include "../../entity/EntityRegistry.h"
 #include "../../entity/Staff.h"
 #include "../../localisation/StringIds.h"
@@ -82,7 +82,7 @@ namespace OpenRCT2::GameActions
             return Result(Status::unknown, STR_CANT_NAME_GUEST, kStringIdNone);
         }
 
-        GfxInvalidateScreen();
+        Drawing::GfxInvalidateScreen();
 
         auto intent = Intent(INTENT_ACTION_REFRESH_STAFF_LIST);
         ContextBroadcastIntent(&intent);
