+++ src/openrct2/actions/ride/RideSetAppearanceAction.cpp
@@ -10,7 +10,7 @@
 #include "RideSetAppearanceAction.h"
 
 #include "../../Diagnostic.h"
-#include "../../drawing/Drawing.h"
+#include "../../drawing/Drawing.Screen.h"
 #include "../../localisation/StringIds.h"
 #include "../../ride/Ride.h"
 #include "../../ui/WindowManager.h"
@@ -100,15 +100,15 @@ namespace OpenRCT2::GameActions
         {
             case RideSetAppearanceType::trackColourMain:
                 ride->trackColours[_index].main = static_cast<Drawing::Colour>(_value);
-                GfxInvalidateScreen();
+                Drawing::GfxInvalidateScreen();
                 break;
             case RideSetAppearanceType::trackColourAdditional:
                 ride->trackColours[_index].additional = static_cast<Drawing::Colour>(_value);
-                GfxInvalidateScreen();
+                Drawing::GfxInvalidateScreen();
                 break;
             case RideSetAppearanceType::trackColourSupports:
                 ride->trackColours[_index].supports = static_cast<Drawing::Colour>(_value);
-                GfxInvalidateScreen();
+                Drawing::GfxInvalidateScreen();
                 break;
             case RideSetAppearanceType::vehicleColourBody:
                 ride->vehicleColours[_index].Body = static_cast<Drawing::Colour>(_value);
@@ -132,7 +132,7 @@ namespace OpenRCT2::GameActions
                 break;
             case RideSetAppearanceType::entranceStyle:
                 ride->entranceStyle = _value;
-                GfxInvalidateScreen();
+                Drawing::GfxInvalidateScreen();
                 break;
             case RideSetAppearanceType::sellingItemColourIsRandom:
                 ride->flags.set(RideFlag::randomShopColours, static_cast<bool>(_value));
