+++ src/openrct2/drawing/Drawing.Screen.h
@@ -1,0 +1,17 @@
+/*****************************************************************************
+ * Copyright (c) 2014-2026 OpenRCT2 developers
+ *
+ * For a complete list of all authors, please refer to contributors.md
+ * Interested in contributing? Visit https://github.com/OpenRCT2/OpenRCT2
+ *
+ * OpenRCT2 is licensed under the GNU General Public License version 3.
+ *****************************************************************************/
+
+#pragma once
+
+namespace OpenRCT2::Drawing
+{
+    void GfxInvalidateScreen();
+    void RefreshVideo();
+    void ToggleWindowedMode();
+} // namespace OpenRCT2::Drawing
