+++ src/openrct2/drawing/Palette.cpp
@@ -17,8 +17,8 @@
 #include "../object/WaterEntry.h"
 #include "../util/Util.h"
 #include "../world/Weather.h"
+#include "Drawing.Screen.h"
 #include "Drawing.Sprite.h"
-#include "Drawing.h"
 #include "FilterPaletteIds.h"
 #include "LightFX.h"
 #include "NewDrawing.h"
