+++ src/openrct2/drawing/PickupPeep.cpp
@@ -8,6 +8,7 @@
 #include "Drawing.Sprite.h"
 #include "Drawing.h"
 #include "ImageId.hpp"
+#include "NewDrawing.h"
 #include "RenderTarget.h"
 
 namespace OpenRCT2::Drawing
