+++ src/openrct2/interface/Chat.cpp
@@ -15,8 +15,8 @@
 #include "../audio/AudioMixer.h"
 #include "../core/UTF8.h"
 #include "../drawing/Drawing.String.h"
-#include "../drawing/Drawing.h"
 #include "../drawing/FilterPaletteIds.h"
+#include "../drawing/NewDrawing.h"
 #include "../drawing/Rectangle.h"
 #include "../drawing/Text.h"
 #include "../interface/ColourWithFlags.h"
